[interface] restructure solver
