+++ include/OGL/Solver/GMRES.hpp
@@ -4,20 +4,18 @@
 
 #pragma once
 
-#include "OGL/GKOlduBase.H"
+#include "OGL/GKOlduBase.hpp"
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
 namespace Foam {
 
-/*---------------------------------------------------------------------------*\
-                           Class GKOBiCGStab Declaration
-\*---------------------------------------------------------------------------*/
-class GKOBiCGStabFactory {
+class GKOGMRESFactory {
 private:
     using mtx = gko::matrix::Csr<scalar>;
     using vec = gko::matrix::Dense<scalar>;
-    using bicgstab = gko::solver::Bicgstab<scalar>;
+    using gmres = gko::solver::Gmres<scalar>;
+    using val_array = gko::array<scalar>;
 
     using dist_vec = gko::experimental::distributed::Vector<scalar>;
     using dist_mtx =
@@ -34,36 +32,34 @@ private:
     mutable std::vector<std::shared_ptr<const gko::stop::CriterionFactory>>
         stoppingCriterionVec_;
 
-
 public:
-    GKOBiCGStabFactory(const dictionary &solverControls,
-                       const objectRegistry &db, word sysMatrixName)
+    GKOGMRESFactory(const dictionary &solverControls, const objectRegistry &db,
+                    word sysMatrixName)
         : solverControls_(solverControls),
           db_(db),
           sysMatrixName_(sysMatrixName),
           stoppingCriterion_(solverControls)
     {}
 
-    CREATE_SOLVER_METHODS(bicgstab)
+    CREATE_SOLVER_METHODS(gmres)
 
-    std::unique_ptr<bicgstab::Factory, std::default_delete<bicgstab::Factory>>
+    std::unique_ptr<gmres::Factory, std::default_delete<gmres::Factory>>
     create_default(std::shared_ptr<gko::Executor> exec) const
     {
-        auto bicgstab = gko::solver::Bicgstab<scalar>::build()
-                            .with_criteria(stoppingCriterionVec_)
-                            .on(exec);
-        return bicgstab;
+        auto gmres =
+            gmres::build().with_criteria(stoppingCriterionVec_).on(exec);
+        return gmres;
     }
 
-    std::unique_ptr<bicgstab::Factory, std::default_delete<bicgstab::Factory>>
+    std::unique_ptr<gmres::Factory, std::default_delete<gmres::Factory>>
     create_precond(std::shared_ptr<gko::Executor> exec,
                    std::shared_ptr<gko::LinOp> precond) const
     {
-        auto bicgstab = bicgstab::build()
-                            .with_criteria(stoppingCriterionVec_)
-                            .with_generated_preconditioner(precond)
-                            .on(exec);
-        return bicgstab;
+        auto gmres = gmres::build()
+                         .with_criteria(stoppingCriterionVec_)
+                         .with_generated_preconditioner(precond)
+                         .on(exec);
+        return gmres;
     }
 
     scalar get_init_res_norm() const
@@ -78,11 +74,6 @@ public:
         return stoppingCriterion_.get_res_norm_time();
     }
 
-    std::shared_ptr<vec> get_res_norms() const
-    {
-        return stoppingCriterion_.get_res_norms();
-    }
-
     scalar get_solve_prev_rel_res_cost() const
     {
         return ::Foam::get_solve_prev_rel_res_cost(sysMatrixName_, db_);
@@ -94,6 +85,11 @@ public:
                                                    prev_rel_res_cost);
     }
 
+
+    std::shared_ptr<vec> get_res_norms() const
+    {
+        return stoppingCriterion_.get_res_norms();
+    }
 
     void store_number_of_iterations() const
     {
@@ -111,37 +107,42 @@ public:
 
     label get_number_of_iterations() const
     {
-        return stoppingCriterion_.get_num_iters() / 2;
+        return stoppingCriterion_.get_num_iters();
     }
 };
 
+/*---------------------------------------------------------------------------*\
+                           Class GKOGMRES Declaration
+\*---------------------------------------------------------------------------*/
 
-class GKOBiCGStab : public GKOlduBaseSolver<GKOBiCGStabFactory> {
+
+class GKOGMRES : public GKOlduBaseSolver<GKOGMRESFactory> {
     // Private Member Functions
 
 public:
-    TypeName("GKOBiCGStab");
+    TypeName("GKOGMRES");
 
     //- Disallow default bitwise copy construct
-    GKOBiCGStab(const GKOBiCGStab &);
+    GKOGMRES(const GKOGMRES &);
 
     //- Disallow default bitwise assignment
-    void operator=(const GKOBiCGStab &);
+    void operator=(const GKOGMRES &);
 
 
     // Constructors
 
     //- Construct from matrix components and solver controls
-    GKOBiCGStab(const word &fieldName, const lduMatrix &matrix,
-                const FieldField<Field, scalar> &interfaceBouCoeffs,
-                const FieldField<Field, scalar> &interfaceIntCoeffs,
-                const lduInterfaceFieldPtrsList &interfaces,
-                const dictionary &solverControls)
+    GKOGMRES(const word &fieldName, const lduMatrix &matrix,
+             const FieldField<Field, scalar> &interfaceBouCoeffs,
+             const FieldField<Field, scalar> &interfaceIntCoeffs,
+             const lduInterfaceFieldPtrsList &interfaces,
+             const dictionary &solverControls)
         : GKOlduBaseSolver(fieldName, matrix, interfaceBouCoeffs,
-                           interfaceIntCoeffs, interfaces, solverControls){};
+                           interfaceIntCoeffs, interfaces, solverControls)
+    {}
 
     //- Destructor
-    virtual ~GKOBiCGStab() {}
+    virtual ~GKOGMRES() {}
 
 
     // Member Functions
