+++ src/Solver/Multigrid.cpp
@@ -3,9 +3,8 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #include <ginkgo/ginkgo.hpp>
-#include <map>
-#include <type_traits>
-#include "GKOMultigrid.H"
+
+#include "OGL/Solver/Multigrid.hpp"
 
 // * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
